Safari-611.1.10/LayoutTests/media/require-user-gesture-to-load-video-expected.txt


EVENT(loadedmetadata)
Did load metadata after user gesture. OK
END OF TEST
